remove the unidentified bonus kind, make Unidentified a prefix
Add generic function for selection from weighted choices.
implement the generic function in gen_adj()

For adjective2 (non-quality adjectives) allow "unidentified", which
entail a second adjective2 selection. Add this to the loot lists.

Add an item_state get_name() function, to hide info when unidentified.
Images on buttons are also hidden, as well as actor_state entries.

Identification at the merchant is added. The first left-click identifies
the item, for 20% of the value, the second click sells the item for the
remaining 80%.

Signed-off-by: picobyte <PIc.O.0ob.y.TE@gmail.com>